#30 Added some improvements to IMDb/Listal Movie sync when sending watched state for rated movies.
It will now send watched for all movies rated (minus whats already watched on trakt). Previously it would only send watched for what it sends as rated (i.e. what is not rated on trakt), this is good but could miss out on some movies if you did not set the option to send watched the first time. It also could lead to duplicates in your watched history if it was already watched.